Moving FAB overrides to utils
